+++ api/services/account/store_createaccount.go
@@ -3,6 +3,7 @@ package account
 import (
 	"context"
 	"errors"
+	"fmt"
 
 	"github.com/MicroFish91/portfolio-instruments-api/api/constants"
 	"github.com/MicroFish91/portfolio-instruments-api/api/types"
@@ -18,10 +19,13 @@ func (s *PostgresAccountStore) CreateAccount(ctx context.Context, a *types.Accou
 
 	row := s.db.QueryRow(
 		c,
-		`INSERT INTO accounts
-		(name, description, tax_shelter, institution, is_deprecated, user_id)
-		VALUES ($1, $2, $3, $4, $5, $6)
-		RETURNING *`,
+		fmt.Sprintf(`
+			insert into accounts
+				(name, description, tax_shelter, institution, is_deprecated, user_id)
+				values ($1, $2, $3, $4, $5, $6)
+			returning
+				%s
+		`, accountColumns),
 		a.Name, a.Description, a.Tax_shelter, a.Institution, a.Is_deprecated, a.User_id,
 	)
