+++ api/services/account/store_getaccountbyid.go
@@ -2,10 +2,10 @@ package account
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/MicroFish91/portfolio-instruments-api/api/constants"
 	"github.com/MicroFish91/portfolio-instruments-api/api/types"
-	"github.com/jackc/pgx/v5"
 )
 
 func (s *PostgresAccountStore) GetAccountById(ctx context.Context, userId int, accountId int) (types.Account, error) {
@@ -14,10 +14,17 @@ func (s *PostgresAccountStore) GetAccountById(ctx context.Context, userId int, a
 
 	row := s.db.QueryRow(
 		c,
-		`SELECT *
-		FROM accounts
-		WHERE user_id = $1 AND account_id = $2`,
-		userId, accountId,
+		fmt.Sprintf(`
+			select
+				%s
+			from
+				accounts
+			where 
+				user_id = $1
+				and account_id = $2
+		`, accountColumns),
+		userId,
+		accountId,
 	)
 
 	account, err := s.parseRowIntoAccount(row)
@@ -26,24 +33,4 @@ func (s *PostgresAccountStore) GetAccountById(ctx context.Context, userId int, a
 	}
 
 	return account, nil
-}
-
-func (s *PostgresAccountStore) parseRowIntoAccount(row pgx.Row) (types.Account, error) {
-	var a types.Account
-	err := row.Scan(
-		&a.Account_id,
-		&a.Name,
-		&a.Description,
-		&a.Tax_shelter,
-		&a.Institution,
-		&a.Is_deprecated,
-		&a.User_id,
-		&a.Created_at,
-		&a.Updated_at,
-	)
-
-	if err != nil {
-		return types.Account{}, err
-	}
-	return a, nil
 }
