+++ api/stores.txt
@@ -12,13 +12,6 @@
 ./api/services/snapshot/store_getsnapshots.go
 ./api/services/snapshot/store_refreshtotal.go
 ./api/services/snapshot/store_groupbyaccount.go
-./api/services/account/store_deleteaccount.go
-./api/services/account/store.go
-./api/services/account/store_getaccounts.go
-./api/services/account/store_getaccountbyid.go
-./api/services/account/store_updateaccount.go
-./api/services/account/store_getaccountbyname.go
-./api/services/account/store_createaccount.go
 ./api/services/snapshotvalue/store.go
 ./api/services/snapshotvalue/store_updatesnapshotvalue.go
 ./api/services/snapshotvalue/store_deletesnapshotvalue.go
